src/main.py
import argparse
import torch
from PIL import Image
from pathlib import Path

def parse_args():
    parser = argparse.ArgumentParser(description='Image Recognition and Text Generation')
    parser.add_argument('--image', type=str, required=True, help='Path to input image')
    return parser.parse_args()

def main():
    args = parse_args()
    
    # Ensure the image exists
    image_path = Path(args.image)
    if not image_path.exists():
        raise FileNotFoundError(f'Image not found: {image_path}')
    
    # Load and process image
    image = Image.open(image_path)
    
    # TODO: Add model loading and inference code
    print(f'Processing image: {image_path}')

if __name__ == '__main__':
    main()