requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.7.0

# Machine Learning
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

# Development Tools
jupyter>=1.0.0
ipython>=8.12.0

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0